+++ NEWS.rst
@@ -13,8 +13,8 @@ The format is based on `Keep a Changelog
 
 Added
 -----
-- Information service collecting album and artist abstract from
-  Wikipedia `#78 <https://github.com/orontee/argos/issues/78>`_
+- Information service collecting abstracts of album and artist pages
+  from Wikipedia `#78 <https://github.com/orontee/argos/issues/78>`_
 
 Changed
 -------